# OpenAPI-schema.yaml
...
openapi: "3.0.3"
info:
  title: "API Schema"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
  version: "0.0.1"
servers:
  - url: "/api"
...
paths:
  /evolve/voicelines/findByName:
    get:
      tags:
        - evolve
      summary: Find the specific character's voice lines by their name
      operationId: EvolveVoiceLinesFindByName
      parameters:
        - $ref: '#/components/parameters/evolveAvailableCharacters'
        - in: query
          name: offset
          required: true
          schema:
            type: integer
            minimum: 0
        - in: query
          name: limit
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 25
        - in: query
          name: sortCriteria
          required: true
          schema:
            $ref: '#/components/schemas/EvolveVoiceLinesSortCriteria'
        - $ref: '#/components/parameters/sortOrder'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvolveVoiceLines'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '5XX':
          $ref: '#/components/responses/UnexpectedError'
...
components:
  schemas:
    EvolveVoiceLine:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        files:
          type: array
          items:
            type: string
      required:
        - id
        - text
    EvolveVoiceLines:
      type: object
      properties:
        lines:
          type: array
          items:
            $ref: '#/components/schemas/EvolveVoiceLine'
      required:
        - lines
    EvolveAvailableCharacters:
      type: string
      enum:
        - markov
        - lennox
        - hyde
        - parnell
        - torvald
        - slim
        - caira
        - lazarus
        - emet
        - val
        - abe
        - scrap
        - griffin
        - crow
        - maggie
        - bucket
        - hank
        - tech
        - sunny
        - cabot
    EvolveVoiceLinesSortCriteria:
      type: string
      enum:
        - id
        - text
    SortOrder:
      type: string
      enum:
        - asc
        - desc
    ResponseError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
...
  parameters:
    sortOrder:
      in: query
      name: sortOrder
      required: true
      schema:
        $ref: '#/components/schemas/SortOrder'
    evolveAvailableCharacters:
      in: query
      name: name
      required: true
      schema:
        $ref: '#/components/schemas/EvolveAvailableCharacters'
  responses:
    BadRequest:
      description: Auto generated text from the request validator
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResponseError'
...
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResponseError'
    UnexpectedError:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResponseError'
...
  securitySchemes:
    petstore_auth:
      type: "oauth2"
      flows:
        implicit:
          authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
          scopes:
            write:pets: "modify pets in your account"
            read:pets: "read your pets"
    api_key:
      type: "apiKey"
      name: "api_key"
      in: "header"